requirements.txt
Django>=3.1.2
psycopg2-binary
djangorestframework
djangorestframework_simplejwt
